refactor: http client types
