+++ packages/auth/src/providers/cognito/utils/clients/HttpClients.ts
@@ -1,20 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import type {
-	ForgotPasswordCommandInput,
-	ForgotPasswordCommandOutput,
-	SignUpCommandInput,
-	SignUpCommandOutput,
-} from '@aws-sdk/client-cognito-identity-provider';
 import { AuthError } from '../../../../errors/AuthError';
 import { assertServiceError } from '../../../../errors/utils/assertServiceError';
+import { ClientOutputs } from './types/outputs';
+import { SendCognitoHttpClientRequest } from './types/request';
 
 const USER_AGENT = 'amplify test';
-
-export type ClientInputs = SignUpCommandInput | ForgotPasswordCommandInput;
-export type ClientOutputs = SignUpCommandOutput | ForgotPasswordCommandOutput;
-export type ClientOperations = 'SignUp' | 'ConfirmSignUp' | 'ForgotPassword';
 
 export class UserPoolHttpClient {
 	private _endpoint: string;
@@ -28,10 +20,10 @@ export class UserPoolHttpClient {
 		this._endpoint = `https://cognito-idp.${region}.amazonaws.com/`;
 	}
 
-	async send<T extends ClientOutputs>(
-		operation: ClientOperations,
-		input: ClientInputs
-	): Promise<T> {
+	async send<T extends ClientOutputs>({
+		operation,
+		input,
+	}: SendCognitoHttpClientRequest): Promise<T> {
 		const headers = {
 			...this._headers,
 			'X-Amz-Target': `AWSCognitoIdentityProviderService.${operation}`,
