+++ packages/auth/src/providers/cognito/utils/clients/types/outputs.ts
@@ -1,0 +1,13 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import type {
+	ForgotPasswordCommandOutput,
+	InitiateAuthCommandOutput,
+	SignUpCommandOutput,
+} from '@aws-sdk/client-cognito-identity-provider';
+
+export type ClientOutputs =
+	| SignUpCommandOutput
+	| ForgotPasswordCommandOutput
+	| InitiateAuthCommandOutput;
